requirements.txt
fastapi
uvicorn
openai
jinja2
aiofiles
requests
python-dotenv
reportlab